refactor: 0회차, 1~8회차, 보너스 미션벌로 제출 섹션 분리화
